abstracted setup into its own class
