+++ src/constants.ts
@@ -382,7 +382,7 @@ export enum TelemetryEventName {
     SETUP_AUTO_RESOLVE_PYTHON_PATH = "SETUP.AUTO.RESOLVE.PYTHON.PATH",
     SETUP_NO_PYTHON_PATH = "SETUP.NO.PYTHON.PATH",
     SETUP_DOWNLOAD_PYTHON = "SETUP.DOWNLOAD.PYTHON",
-    SETUP_BAD_PYTHON_PATH = "SETUP.BAD.PYTHON.PATH",
+    SETUP_INVALID_PYTHON_INTERPRETER_PATH = "SETUP.INVALID.PYTHON.INTERPRETER.PATH",
     SETUP_INVALID_PYTHON_VER = "SETUP.INVALID.PYTHON.VER",
     SETUP_INSTALL_VENV = "SETUP.INSTALL.VENV",
     SETUP_ORIGINAL_INTERPRETER_DEP_INSTALL = "SETUP.ORIGINAL.INTERPRETER.DEP.INSTALL",
