+++ src/extension.ts
@@ -27,6 +27,7 @@ import { SerialMonitor } from "./serialMonitor";
 import { DebuggerCommunicationService } from "./service/debuggerCommunicationService";
 import { MessagingService } from "./service/messagingService";
 import { PopupService } from "./service/PopupService";
+import { SetupService } from "./service/SetupService";
 import { SimulatorDebugConfigurationProvider } from "./simulatorDebugConfigurationProvider";
 import getPackageInfo from "./telemetry/getPackageInfo";
 import TelemetryAI from "./telemetry/telemetryAI";
@@ -43,7 +44,9 @@ let inDebugMode: boolean = false;
 let firstTimeClosed: boolean = true;
 let shouldShowRunCodePopup: boolean = true;
 const messagingService = new MessagingService();
+let setupService: SetupService;
 const debuggerCommunicationService = new DebuggerCommunicationService();
+
 
 let currentActiveDevice: string = DEFAULT_DEVICE;
 
@@ -87,6 +90,7 @@ export async function activate(context: vscode.ExtensionContext) {
     console.info(CONSTANTS.INFO.EXTENSION_ACTIVATED);
 
     telemetryAI = new TelemetryAI(context);
+    setupService = new SetupService(telemetryAI);
     let currentPanel: vscode.WebviewPanel | undefined;
     let childProcess: cp.ChildProcess | undefined;
     let messageListener: vscode.Disposable;
@@ -99,7 +103,7 @@ export async function activate(context: vscode.ExtensionContext) {
     // doesn't trigger lint errors
     updatePylintArgs(context);
 
-    pythonExecutablePath = await utils.setupEnv(context, telemetryAI);
+    pythonExecutablePath = await setupService.setupEnv(context, telemetryAI);
 
     try {
         utils.generateCPXConfig();
@@ -443,7 +447,7 @@ export async function activate(context: vscode.ExtensionContext) {
     const installDependencies: vscode.Disposable = vscode.commands.registerCommand(
         "deviceSimulatorExpress.common.installDependencies",
         async () => {
-            pythonExecutablePath = await utils.setupEnv(
+            pythonExecutablePath = await setupService.setupEnv(
                 context,
                 telemetryAI,
                 true
@@ -1032,7 +1036,7 @@ export async function activate(context: vscode.ExtensionContext) {
     const configsChanged = vscode.workspace.onDidChangeConfiguration(
         async () => {
             if (utils.checkConfig(CONFIG.CONFIG_ENV_ON_SWITCH)) {
-                pythonExecutablePath = await utils.setupEnv(
+                pythonExecutablePath = await setupService.setupEnv(
                     context,
                     telemetryAI
                 );
@@ -1087,7 +1091,7 @@ const updateCurrentFileIfPython = async (
     if (
         currentTextDocument &&
         utils.getActiveEditorFromPath(currentTextDocument.fileName) ===
-            undefined
+        undefined
     ) {
         await vscode.window.showTextDocument(
             currentTextDocument,
