+++ src/extension_utils/utils.ts
@@ -3,27 +3,20 @@
 
 import * as cp from "child_process";
 import * as fs from "fs";
-import * as open from "open";
-import * as os from "os";
 import * as path from "path";
 import * as util from "util";
 import * as vscode from "vscode";
 import {
-    CONFIG,
     CONSTANTS,
     CPX_CONFIG_FILE,
     DialogResponses,
     GLOBAL_ENV_VARS,
-    HELPER_FILES,
     SERVER_INFO,
-    TelemetryEventName,
-    VERSIONS,
 } from "../constants";
 import { CPXWorkspace } from "../cpxWorkspace";
 import { DeviceContext } from "../deviceContext";
-import TelemetryAI from "../telemetry/telemetryAI";
 
-const exec = util.promisify(cp.exec);
+export const exec = util.promisify(cp.exec);
 
 const errorChannel = vscode.window.createOutputChannel(
     CONSTANTS.ERROR.INSTALLATION_ERROR
@@ -80,7 +73,7 @@ export function tryParseJSON(jsonString: string): any | boolean {
         if (jsonObj && typeof jsonObj === "object") {
             return jsonObj;
         }
-    } catch (exception) {}
+    } catch (exception) { }
 
     return false;
 }
@@ -164,34 +157,6 @@ export function generateCPXConfig(): void {
     fs.writeFileSync(cpxConfigFilePath, JSON.stringify(cpxJson, null, 4));
 }
 
-export const isPipInstalled = async (pythonExecutablePath: string) => {
-    try {
-        const { stdout } = await executePythonCommand(
-            pythonExecutablePath,
-            " -m pip"
-        );
-        return true;
-    } catch (err) {
-        vscode.window
-            .showErrorMessage(
-                CONSTANTS.ERROR.NO_PIP,
-                DialogResponses.INSTALL_PIP
-            )
-            .then((selection: vscode.MessageItem | undefined) => {
-                if (selection === DialogResponses.INSTALL_PIP) {
-                    const okAction = () => {
-                        open(CONSTANTS.LINKS.DOWNLOAD_PIP);
-                    };
-                    showPrivacyModal(
-                        okAction,
-                        CONSTANTS.INFO.THIRD_PARTY_WEBSITE_PIP
-                    );
-                }
-            });
-        return false;
-    }
-};
-
 export const addVisibleTextEditorCallback = (
     currentPanel: vscode.WebviewPanel,
     context: vscode.ExtensionContext
@@ -265,446 +230,4 @@ export const getTelemetryState = () => {
     return vscode.workspace
         .getConfiguration()
         .get("telemetry.enableTelemetry", true);
-};
-
-// Setup code starts
-
-export const checkIfVenv = async (
-    context: vscode.ExtensionContext,
-    pythonExecutablePath: string
-) => {
-    const venvCheckerPath: string = getPathToScript(
-        context,
-        CONSTANTS.FILESYSTEM.OUTPUT_DIRECTORY,
-        HELPER_FILES.CHECK_IF_VENV_PY
-    );
-    const { stdout } = await executePythonCommand(
-        pythonExecutablePath,
-        `"${venvCheckerPath}"`
-    );
-    return stdout.trim() === "1";
-};
-
-export const executePythonCommand = async (
-    pythonExecutablePath: string,
-    command: string
-) => {
-    return exec(`${createEscapedPath(pythonExecutablePath)} ${command}`);
-};
-
-export const validatePythonVersion = async (pythonExecutablePath: string) => {
-    const { stdout } = await executePythonCommand(
-        pythonExecutablePath,
-        "--version"
-    );
-    if (stdout < VERSIONS.MIN_PY_VERSION) {
-        vscode.window
-            .showInformationMessage(
-                CONSTANTS.ERROR.INVALID_PYTHON_PATH,
-                DialogResponses.INSTALL_PYTHON
-            )
-            .then((installChoice: vscode.MessageItem | undefined) => {
-                if (installChoice === DialogResponses.INSTALL_PYTHON) {
-                    const okAction = () => {
-                        open(CONSTANTS.LINKS.DOWNLOAD_PYTHON);
-                    };
-                    showPrivacyModal(
-                        okAction,
-                        CONSTANTS.INFO.THIRD_PARTY_WEBSITE_PYTHON
-                    );
-                }
-            });
-        return false;
-    } else {
-        return true;
-    }
-};
-
-export const hasVenv = async (context: vscode.ExtensionContext) => {
-    const pathToEnv: string = getPathToScript(
-        context,
-        CONSTANTS.FILESYSTEM.PYTHON_VENV_DIR
-    );
-
-    return fs.existsSync(pathToEnv);
-};
-
-export const promptInstallVenv = (
-    context: vscode.ExtensionContext,
-    pythonExecutable: string,
-    pythonExecutableName: string,
-    telemetryAI: TelemetryAI
-) => {
-    return vscode.window
-        .showInformationMessage(
-            CONSTANTS.INFO.INSTALL_PYTHON_VENV,
-            DialogResponses.YES,
-            DialogResponses.NO
-        )
-        .then((selection: vscode.MessageItem | undefined) => {
-            if (selection === DialogResponses.YES) {
-                return installPythonVenv(
-                    context,
-                    pythonExecutable,
-                    pythonExecutableName,
-                    telemetryAI
-                );
-            } else {
-                // return pythonExecutable, notifying the caller
-                // that the user was unwilling to create venv
-                // and by default, this will trigger the extension to
-                // try using pythonExecutable
-                return pythonExecutable;
-            }
-        });
-};
-
-export const getPythonVenv = async (
-    context: vscode.ExtensionContext,
-    pythonExecutableName: string
-) => {
-    const subFolder = os.platform() === "win32" ? "Scripts" : "bin";
-
-    return getPathToScript(
-        context,
-        path.join(CONSTANTS.FILESYSTEM.PYTHON_VENV_DIR, subFolder),
-        pythonExecutableName
-    );
-};
-
-export const installPythonVenv = async (
-    context: vscode.ExtensionContext,
-    pythonExecutable: string,
-    pythonExecutableName: string,
-    telemetryAI: TelemetryAI
-) => {
-    const pathToEnv: string = getPathToScript(
-        context,
-        CONSTANTS.FILESYSTEM.PYTHON_VENV_DIR
-    );
-
-    vscode.window.showInformationMessage(CONSTANTS.INFO.INSTALLING_PYTHON_VENV);
-
-    const pythonPath: string = await getPythonVenv(
-        context,
-        pythonExecutableName
-    );
-
-    try {
-        // make venv
-        // run command to download dependencies to out/python_libs
-        await executePythonCommand(pythonExecutable, `-m venv "${pathToEnv}"`);
-    } catch (err) {
-        telemetryAI.trackFeatureUsage(
-            TelemetryEventName.SETUP_VENV_CREATION_ERR
-        );
-        vscode.window
-            .showErrorMessage(
-                `Virtual environment for download could not be completed. Using original interpreter at: ${pythonExecutable}.`,
-                DialogResponses.READ_INSTALL_MD
-            )
-            .then((selection: vscode.MessageItem | undefined) => {
-                if (selection === DialogResponses.READ_INSTALL_MD) {
-                    open(CONSTANTS.LINKS.INSTALL);
-                }
-            });
-
-        console.error(err);
-
-        return pythonExecutable;
-    }
-
-    return installDependenciesWrapper(
-        context,
-        pythonPath,
-        telemetryAI,
-        pythonExecutable
-    );
-};
-
-export const areDependenciesInstalled = async (
-    context: vscode.ExtensionContext,
-    pythonPath: string
-) => {
-    const dependencyCheckerPath: string = getPathToScript(
-        context,
-        CONSTANTS.FILESYSTEM.OUTPUT_DIRECTORY,
-        HELPER_FILES.CHECK_PYTHON_DEPENDENCIES
-    );
-    try {
-        // python script will throw exception
-        // if not all dependencies are downloaded
-        const { stdout } = await executePythonCommand(
-            pythonPath,
-            `"${dependencyCheckerPath}"`
-        );
-
-        // output for debugging purposes
-        console.info(stdout);
-        return true;
-    } catch (err) {
-        return false;
-    }
-};
-
-export const installDependencies = async (
-    context: vscode.ExtensionContext,
-    pythonPath: string,
-    telemetryAI: TelemetryAI
-) => {
-    const requirementsPath: string = getPathToScript(
-        context,
-        CONSTANTS.FILESYSTEM.OUTPUT_DIRECTORY,
-        "requirements.txt"
-    );
-
-    if (!isPipInstalled(pythonPath)) {
-        telemetryAI.trackFeatureUsage(TelemetryEventName.SETUP_NO_PIP);
-        return false;
-    }
-
-    try {
-        const { stdout } = await executePythonCommand(
-            pythonPath,
-            `-m pip install -r "${requirementsPath}"`
-        );
-
-        console.info(stdout);
-        vscode.window.showInformationMessage(CONSTANTS.INFO.SUCCESSFUL_INSTALL);
-        return true;
-    } catch (err) {
-        return false;
-    }
-};
-
-export const installDependenciesWrapper = async (
-    context: vscode.ExtensionContext,
-    pythonPath: string,
-    telemetryAI: TelemetryAI,
-    backupPythonPath: string = ""
-) => {
-    let errMessage = CONSTANTS.ERROR.DEPENDENCY_DOWNLOAD_ERROR;
-    if (backupPythonPath !== "") {
-        errMessage = `${errMessage} Using original interpreter at: ${backupPythonPath}.`;
-    }
-    if (!(await installDependencies(context, pythonPath, telemetryAI))) {
-        vscode.window
-            .showErrorMessage(
-                CONSTANTS.ERROR.DEPENDENCY_DOWNLOAD_ERROR,
-                DialogResponses.READ_INSTALL_MD
-            )
-            .then((selection: vscode.MessageItem | undefined) => {
-                if (selection === DialogResponses.READ_INSTALL_MD) {
-                    open(CONSTANTS.LINKS.INSTALL);
-                }
-            });
-
-        telemetryAI.trackFeatureUsage(
-            TelemetryEventName.SETUP_DEP_INSTALL_FAIL
-        );
-        return backupPythonPath;
-    }
-    return pythonPath;
-};
-export const getCurrentpythonExecutablePath = async (
-    telemetryAI: TelemetryAI
-) => {
-    let originalpythonExecutablePath = "";
-
-    // try to get name from interpreter
-    try {
-        originalpythonExecutablePath = getConfig(CONFIG.PYTHON_PATH);
-    } catch (err) {
-        originalpythonExecutablePath = GLOBAL_ENV_VARS.PYTHON;
-    }
-
-    if (
-        originalpythonExecutablePath === GLOBAL_ENV_VARS.PYTHON ||
-        originalpythonExecutablePath === ""
-    ) {
-        telemetryAI.trackFeatureUsage(
-            TelemetryEventName.SETUP_AUTO_RESOLVE_PYTHON_PATH
-        );
-        try {
-            const { stdout } = await executePythonCommand(
-                GLOBAL_ENV_VARS.PYTHON,
-                `-c "import sys; print(sys.executable)"`
-            );
-            originalpythonExecutablePath = stdout.trim();
-        } catch (err) {
-            telemetryAI.trackFeatureUsage(
-                TelemetryEventName.SETUP_NO_PYTHON_PATH
-            );
-            vscode.window
-                .showErrorMessage(
-                    CONSTANTS.ERROR.NO_PYTHON_PATH,
-                    DialogResponses.INSTALL_PYTHON
-                )
-                .then((selection: vscode.MessageItem | undefined) => {
-                    if (selection === DialogResponses.INSTALL_PYTHON) {
-                        const okAction = () => {
-                            telemetryAI.trackFeatureUsage(
-                                TelemetryEventName.SETUP_DOWNLOAD_PYTHON
-                            );
-                            open(CONSTANTS.LINKS.DOWNLOAD_PYTHON);
-                        };
-                        showPrivacyModal(
-                            okAction,
-                            CONSTANTS.INFO.THIRD_PARTY_WEBSITE_PYTHON
-                        );
-                    }
-                });
-            // no python installed, cannot get path
-            return "";
-        }
-    }
-    // fix path to be absolute
-    if (!path.isAbsolute(originalpythonExecutablePath)) {
-        originalpythonExecutablePath = path.join(
-            vscode.workspace.rootPath,
-            originalpythonExecutablePath
-        );
-    }
-
-    if (!fs.existsSync(originalpythonExecutablePath)) {
-        await vscode.window.showErrorMessage(CONSTANTS.ERROR.BAD_PYTHON_PATH);
-        telemetryAI.trackFeatureUsage(TelemetryEventName.SETUP_BAD_PYTHON_PATH);
-        return "";
-    }
-
-    if (!(await validatePythonVersion(originalpythonExecutablePath))) {
-        telemetryAI.trackFeatureUsage(
-            TelemetryEventName.SETUP_INVALID_PYTHON_VER
-        );
-        return "";
-    }
-
-    return originalpythonExecutablePath;
-};
-export const setupEnv = async (
-    context: vscode.ExtensionContext,
-    telemetryAI: TelemetryAI,
-    needsResponse: boolean = false
-) => {
-    const originalpythonExecutablePath = await getCurrentpythonExecutablePath(
-        telemetryAI
-    );
-    let pythonExecutablePath = originalpythonExecutablePath;
-    const pythonExecutableName: string =
-        os.platform() === "win32"
-            ? HELPER_FILES.PYTHON_EXE
-            : HELPER_FILES.PYTHON;
-
-    if (!(await areDependenciesInstalled(context, pythonExecutablePath))) {
-        // environment needs to install dependencies
-        if (!(await checkIfVenv(context, pythonExecutablePath))) {
-            const pythonExecutablePathVenv = await getPythonVenv(
-                context,
-                pythonExecutableName
-            );
-            if (await hasVenv(context)) {
-                // venv in extention exists with wrong dependencies
-
-                if (
-                    !(await areDependenciesInstalled(
-                        context,
-                        pythonExecutablePathVenv
-                    ))
-                ) {
-                    pythonExecutablePath = await installDependenciesWrapper(
-                        context,
-                        pythonExecutablePathVenv,
-                        telemetryAI,
-                        pythonExecutablePath
-                    );
-                } else {
-                    pythonExecutablePath = pythonExecutablePathVenv;
-                }
-            } else {
-                pythonExecutablePath = await promptInstallVenv(
-                    context,
-                    originalpythonExecutablePath,
-                    pythonExecutableName,
-                    telemetryAI
-                );
-                telemetryAI.trackFeatureUsage(
-                    TelemetryEventName.SETUP_INSTALL_VENV
-                );
-            }
-
-            if (pythonExecutablePath === pythonExecutablePathVenv) {
-                vscode.window.showInformationMessage(
-                    CONSTANTS.INFO.UPDATED_TO_EXTENSION_VENV
-                );
-                vscode.workspace
-                    .getConfiguration()
-                    .update(CONFIG.PYTHON_PATH, pythonExecutablePath);
-            }
-        } else {
-            telemetryAI.trackFeatureUsage(TelemetryEventName.SETUP_HAS_VENV);
-        }
-        if (pythonExecutablePath === originalpythonExecutablePath) {
-            // going with original interpreter, either because
-            // already in venv or error in creating custom venv
-            if (checkConfig(CONFIG.SHOW_DEPENDENCY_INSTALL)) {
-                await vscode.window
-                    .showInformationMessage(
-                        CONSTANTS.INFO.INSTALL_PYTHON_DEPS,
-                        DialogResponses.INSTALL_NOW,
-                        DialogResponses.DONT_INSTALL
-                    )
-                    .then(
-                        async (
-                            installChoice: vscode.MessageItem | undefined
-                        ) => {
-                            if (installChoice === DialogResponses.INSTALL_NOW) {
-                                telemetryAI.trackFeatureUsage(
-                                    TelemetryEventName.SETUP_ORIGINAL_INTERPRETER_DEP_INSTALL
-                                );
-                                await installDependenciesWrapper(
-                                    context,
-                                    pythonExecutablePath,
-                                    telemetryAI
-                                );
-                            } else {
-                                await vscode.window
-                                    .showInformationMessage(
-                                        CONSTANTS.INFO.ARE_YOU_SURE,
-                                        DialogResponses.INSTALL_NOW,
-                                        DialogResponses.DONT_INSTALL
-                                    )
-                                    .then(
-                                        async (
-                                            installChoice2:
-                                                | vscode.MessageItem
-                                                | undefined
-                                        ) => {
-                                            if (
-                                                installChoice2 ===
-                                                DialogResponses.INSTALL_NOW
-                                            ) {
-                                                await installDependenciesWrapper(
-                                                    context,
-                                                    pythonExecutablePath,
-                                                    telemetryAI
-                                                );
-                                            } else {
-                                                telemetryAI.trackFeatureUsage(
-                                                    TelemetryEventName.SETUP_NO_DEPS_INSTALLED
-                                                );
-                                            }
-                                        }
-                                    );
-                            }
-                        }
-                    );
-            }
-        }
-    } else if (needsResponse) {
-        vscode.window.showInformationMessage(
-            CONSTANTS.INFO.ALREADY_SUCCESSFUL_INSTALL
-        );
-    }
-
-    return pythonExecutablePath;
 };
